+++ build.gradle
@@ -56,7 +56,6 @@ ext {
     espressoCoreVersion = "3.0.2"
     jsonsnapshotVersion = "1.0.12"
     androidxTestCoreVersion = "1.0.0"
-    espressoIntentsVersion = "3.0.1"
 }
 
 task buildci(type: GradleBuild) {
