--- example/src/androidTest/res/raw/preset.json
@@ -1,28 +1,0 @@
-{
-    "transactionId": "tr200",
-    "integration": "DISPLAY_NATIVE",
-    "presetFirst": "false",
-    "country": "DE",
-    "customer": {
-        "number": "42",
-        "email": "john.doe@example.com",
-        "registration": {
-            "id": "5c51c081b55f286f2db9f97au"
-        }
-    },
-    "payment": {
-        "amount": 9.99,
-        "currency": "EUR",
-        "reference": "Shop optile/03-12-2018"
-    },
-    "style": {
-        "hostedVersion": "v2",
-        "language": "en_US"
-    },
-    "callback": {
-        "returnUrl": "https://dev.oscato.com/shop/success.html",
-        "summaryUrl": "https://dev.oscato.com/shop/summary.html",
-        "cancelUrl": "https://dev.oscato.com/shop/cancel.html",
-        "notificationUrl": "https://dev.oscato.com/shop/notify.html"
-    }
-}
