+++ payment/src/main/java/net/optile/payment/network/ImageConnection.java
@@ -6,7 +6,7 @@
  * See the LICENSE file for more information.
  */
 
-package net.optile.payment.list;
+package net.optile.payment.network;
 
 import static net.optile.payment.core.PaymentError.CONN_ERROR;
