+++ payment/src/main/java/net/optile/payment/ui/page/PaymentPageService.java
@@ -32,8 +32,8 @@ import net.optile.payment.model.ListResult;
 import net.optile.payment.model.Networks;
 import net.optile.payment.model.OperationResult;
 import net.optile.payment.model.PresetAccount;
-import net.optile.payment.list.ListConnection;
-import net.optile.payment.list.PaymentConnection;
+import net.optile.payment.network.ListConnection;
+import net.optile.payment.network.PaymentConnection;
 import net.optile.payment.resource.PaymentGroup;
 import net.optile.payment.resource.ResourceLoader;
 import net.optile.payment.ui.PaymentUI;
