+++ payment/src/test/java/net/optile/payment/network/ListConnectionTest.java
@@ -6,7 +6,7 @@
  * See the LICENSE file for more information.
  */
 
-package net.optile.payment.list;
+package net.optile.payment.network;
 
 import static org.junit.Assert.assertNotNull;
